+++ CHANGELOG.md
@@ -16,6 +16,8 @@ CHANGELOG for FlatCAM Evo beta
 - updated the hack to make sure that the Editor sub-tools do not lose the stylesheet of the background
 - updated the disabled project item color default value for the dark theme
 - in Document Object UI - upgraded the UI and fixed the colors setting
+- some refactoring and updated the application strings
+- updated Romanian translation
 
 10.05.2022